move troposphere-specific interpolation to its subpackage
