+++ src/dolphin/__init__.py
@@ -5,5 +5,3 @@ from dolphin._version import version as __version__
 from ._log import *
 from ._show_versions import *
 from ._types import *
-from .goldstein import goldstein
-from .interpolation import interpolate
